+++ crates/code_assistant/src/tests/mocks.rs
