+++ crates/code_assistant/src/tests/mod.rs
@@ -1,0 +1,1 @@
+pub mod mocks;
