+++ crates/code_assistant/src/tools/impls/delete_files.rs
@@ -172,7 +172,7 @@ impl Tool for DeleteFilesTool {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::tools::tests::mocks::{MockExplorer, MockProjectManager};
+    use crate::tests::mocks::{MockExplorer, MockProjectManager};
     use crate::types::WorkingMemory;
     use std::collections::HashMap;
