+++ crates/code_assistant/src/tools/impls/read_files.rs
@@ -222,6 +222,9 @@ impl Tool for ReadFilesTool {
 #[cfg(test)]
 mod tests {
     use super::*;
+    use crate::tests::mocks::{MockExplorer, MockProjectManager};
+    use crate::tools::core::ToolRegistry;
+    use crate::types::WorkingMemory;
 
     #[tokio::test]
     async fn test_read_files_output_rendering() {
@@ -246,5 +249,73 @@ mod tests {
         assert!(rendered.contains("File not found"));
         assert!(rendered.contains(">>>>> FILE: test.txt"));
         assert!(rendered.contains("Test file content"));
+    }
+
+    #[tokio::test]
+    async fn test_read_files_updates_memory() -> Result<()> {
+        // Create a tool registry
+        let registry = ToolRegistry::global();
+
+        // Get the read_files tool
+        let read_files_tool = registry
+            .get("read_files")
+            .expect("read_files tool should be registered");
+
+        // Create test files
+        let mut files = HashMap::new();
+        files.insert(
+            PathBuf::from("./root/test.txt"),
+            "Test file content".to_string(),
+        );
+        files.insert(
+            PathBuf::from("./root/test2.txt"),
+            "Another file content".to_string(),
+        );
+
+        // Create a mock explorer with these files
+        let explorer = MockExplorer::new(files, None);
+
+        // Create working memory
+        let mut working_memory = WorkingMemory::default();
+
+        // Create a tool context with working memory
+        let mut context = ToolContext::<'_> {
+            project_manager: &MockProjectManager::default().with_project(
+                "test-project",
+                PathBuf::from("./root"),
+                explorer,
+            ),
+            command_executor: &crate::utils::DefaultCommandExecutor,
+            working_memory: Some(&mut working_memory),
+        };
+
+        // Parameters for read_files
+        let params = json!({
+            "project": "test-project",
+            "paths": ["test.txt", "test2.txt"]
+        });
+
+        // Execute the tool
+        let result = read_files_tool.invoke(&mut context, params).await?;
+
+        // Format the output
+        let mut tracker = crate::tools::core::ResourcesTracker::new();
+        let output = result.as_render().render(&mut tracker);
+
+        // Check the output
+        assert!(output.contains("Successfully loaded"));
+
+        // Verify that the files were added to working memory
+        assert_eq!(working_memory.loaded_resources.len(), 2);
+
+        // Check that both files are in the working memory
+        assert!(working_memory
+            .loaded_resources
+            .contains_key(&("test-project".to_string(), "test.txt".into())));
+        assert!(working_memory
+            .loaded_resources
+            .contains_key(&("test-project".to_string(), "test2.txt".into())));
+
+        Ok(())
     }
 }
