--- crates/code_assistant/src/tools/tests/mod.rs
@@ -1,4 +1,0 @@
-// Integration tests for tool system
-pub mod integration_test;
-pub mod mocks;
-mod working_memory_test;
